+++ web/concrete/src/Utility/UtilityServiceProvider.php
@@ -6,7 +6,6 @@ class UtilityServiceProvider extends ServiceProvider {
 	public function register() {
 		$singletons = array(
             'helper/text' => '\Concrete\Core\Utility\Service\Text',
-            'helper/html' => '\Concrete\Core\Utility\Service\Html',
 			'helper/arrays' => '\Concrete\Core\Utility\Service\Arrays',
 			'helper/number' => '\Concrete\Core\Utility\Service\Number',
 			'helper/xml' => '\Concrete\Core\Utility\Service\Xml',
